// src/app/ServiceSwagger/model/postDto.ts
/**
 * Backend
 * Descripción
 *
 * OpenAPI spec version: 0.0.1
 * Contact: willypanza11@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Usuario } from './usuario';

export interface PostDto { 
    content?: string;
    createdOn?: Date;
    id?: number;
    iduser?: number;
    title?: string;
    updatedOn?: Date;
    user?: Usuario;
    username?: string;
}